refactor(arch): Phase 3 — infrastructure/ai → entities/llm-provider, infrastructure/llm → entities/api-key/actions
- AI provider 6개 파일 → entities/llm-provider/{api,lib}
- LLM Server Action 3개 → entities/api-key/actions/
- 테스트 9개 colocated 이동
- entities/llm-provider/index.ts barrel 생성
- 12개 파일 import 갱신, 내부 relative import 전환
- infrastructure/ai/, infrastructure/llm/ 디렉토리 제거

Phase 3 of FSD migration (PR A: AI + LLM)
