+++ src/entities/api-key/__tests__/actions/saveApiKeyAction.test.ts
@@ -23,7 +23,7 @@ jest.mock('@/entities/api-key', () => ({
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
-import { saveApiKeyAction } from '@/infrastructure/llm/saveApiKeyAction';
+import { saveApiKeyAction } from '@/entities/api-key/actions/saveApiKeyAction';
 import { makeFormData } from '@/__tests__/utils/makeFormData';
 import type { ApiKeyActionState } from '@/domain/llm';
