+++ src/entities/llm-provider/__tests__/lib/parseJsonResponse.test.ts
@@ -1,7 +1,7 @@
 import {
     parseJsonResponse,
     stripMarkdownCodeBlock,
-} from '@/infrastructure/ai/parseJsonResponse';
+} from '@/entities/llm-provider/lib/parseJsonResponse';
 
 describe('stripMarkdownCodeBlock', () => {
     it('strips ```json fenced wrapper and trims inner content', () => {
