+++ src/entities/llm-provider/lib/utils.ts
