+++ src/infrastructure/auth/errorMessages.ts
@@ -1,4 +1,4 @@
-// Korean-only by design (single-locale product); future i18n keys map 1:1 to the constants in this file. Sibling: lib/contactErrorMessages.ts, infrastructure/llm/saveApiKeyAction.ts.
+// Korean-only by design (single-locale product); future i18n keys map 1:1 to the constants in this file. Sibling: lib/contactErrorMessages.ts, entities/api-key/actions/saveApiKeyAction.ts.
 
 export const AUTH_SERVICE_UNAVAILABLE_MESSAGE =
     '서비스가 일시적으로 동작하지 않습니다. 잠시 후 다시 시도해주세요.';
