+++ src/infrastructure/chat/chatAction.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { callAiProviderRouter } from '@/infrastructure/ai/router';
+import { callAiProviderRouter } from '@/entities/llm-provider';
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { getDatabaseClient } from '@/shared/db/client';
 import { DrizzleUserApiKeyRepository } from '@/entities/api-key';
